requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-unfold
drf-spectacular
django-environ
django-constance
psycopg2-binary
celery[redis]
django_celery_beat
pillow
cryptography
django-filter
python-dateutil
requests
simple-salesforce
django-ckeditor-5
gunicorn==21.2.0
django-redis
factory_boy
Faker
django-cors-headers
openpyxl
reportlab